Internals

BangBang.Experimental.mergewith!! — Function

mergewith!!(combine, dict, others...) -> dict′
mergewith!!(combine)

Like merge!!(combine, dict, others...) but combine does not have to be a Function.

The curried form mergewith!!(combine) returns the function (args...) -> mergewith!!(combine, args...).

See also: Add mergewith[!](combine, dicts...) by tkf · Pull Request #34296 · JuliaLang/julia

BangBang.Experimental.modify!! — Function

modify!!(f, dictlike, key) -> (dictlike′, ret)

Lookup and then update, insert or delete in one go. If supported (e.g., when dictlike isa Dict), it is done without re-computing the hash. Immutable containers like NamedTuple is also supported.

The callable f must accept a single argument of type Union{Some{valtype(dictlike)}, Nothing}. The value Some(dictlike[key]) is passed to f if haskey(dictlike, key); otherwise nothing is passed.

If f returns nothing, corresponding entry in the dictionary dictlike is removed. If f returns non-nothing value x, key => something(x) is inserted to dictlike (equivalent to dictlike[key] = something(x) but more efficient).

modify!! returns a 2-tuple (dictlike′, ret) where dictlike′ is an updated version of dictlike (which may be identical to dictlike) and ret is the returned value of f.

This API is inspired by Control.Lens.At of Haskell's lens library.

Examples

julia> using BangBang.Experimental: modify!!

julia> dict = Dict("a" => 1);

julia> dict′, ret = modify!!(dict, "a") do val
           Some(something(val, 0) + 1)
       end;

julia> ret
Some(2)

julia> dict === dict′
true

julia> dict
Dict{String,Int64} with 1 entry:
  "a" => 2

julia> dict = Dict();

julia> modify!!(dict, "a") do val
           Some(something(val, 0) + 1)
       end;

julia> dict
Dict{Any,Any} with 1 entry:
  "a" => 1

julia> modify!!(_ -> nothing, dict, "a");

julia> dict
Dict{Any,Any} with 0 entries

Discussion

Add modify! function for lookup/update/insert/delete in one go by tkf · Pull Request #33758 · JuliaLang/julia

Support adding to named tuples? · Issue #65 · jw3126/Setfield.jl

BangBang.implements — Method

implements(f!, x) :: Bool

true if in-place function f! can mutate x.

Examples

julia> using BangBang: implements

julia> implements(push!, Vector)
true

julia> implements(push!, [])  # works with instances, too
true

julia> implements(push!, Tuple)
false

julia> using StaticArrays

julia> implements(push!, SVector)
false

julia> implements(setindex!, SVector)
false

julia> implements(push!, MVector)
false

julia> implements(setindex!, MVector)
true

BangBang.may — Method

may(mutate!, args...)

Call mutate!(args...) if possible; fallback to the out-of-place version if not.

BangBang.possible — Function

possible(f!, args...) :: Bool

Check if f!(args...) is possible.

Examples

julia> using BangBang: possible

julia> possible(push!, Int[], 1)
true

julia> possible(push!, Int[], 0.5)
false

BangBang.resize!! — Method

resize!!(vector::AbstractVector, n::Integer) -> vector′